[#3] Fix indendt and dedent tokens not generating
- Moved indentation algorithm variables definition from header
	file to *.c file as it caused algorithm not to work
	properly (as there were multiple definitions)
