+++ PLT_Parser/PLT_Parser/Grammars/PyToAsm/SyntaxMainHeader.h
@@ -1,3 +1,9 @@
 #pragma once
 
+#ifdef __cplusplus
+extern "C" {
+#endif
 int parser_main(int, char**);
+#ifdef __cplusplus
+}
+#endif
