+++ PLT_Parser/PLT_Parser/Tests/PyToAsm/test1.txt
@@ -1,0 +1,8 @@
+
+if a <= 3:
+	b = 3
+
+def main():
+	a = 3
+	while a:
+		return
