+++ PLT_Parser/PLT_Parser/x64/Debug/PLT_Parser.Build.CppClean.log
@@ -1,22 +1,18 @@
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\parser.tab.h
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\parser.tab.c
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\parser.tab.cpp
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\lexanalyzer.yy.c
+d:\projects\c#\plangtranslator\plt_parser\plt_parser\grammars\pytoasm\syntaxanal.tab.cpp
+d:\projects\c#\plangtranslator\plt_parser\plt_parser\grammars\pytoasm\lexanal.flex.cpp
 d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\vc143.pdb
 d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\vc143.idb
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\lexalgorithms.obj
 d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\main.obj
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.write.1.tlog
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\tools.write.1.tlog
+d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\syntaxanal.tab.obj
+d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\lexanal.flex.obj
+d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\lexalgorithms.obj
+d:\projects\c#\plangtranslator\plt_parser\x64\debug\plt_parser.exe
+d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.ilk
 d:\projects\c#\plangtranslator\plt_parser\x64\debug\plt_parser.pdb
+d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.write.1.tlog
 d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\cl.command.1.tlog
 d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\cl.read.1.tlog
 d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\cl.write.1.tlog
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\link-cvtres.read.1.tlog
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\link-cvtres.write.1.tlog
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\link-rc.read.1.tlog
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\link-rc.write.1.tlog
 d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\link.command.1.tlog
 d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\link.read.1.tlog
-d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\link.read.2.tlog
 d:\projects\c#\plangtranslator\plt_parser\plt_parser\x64\debug\plt_parser.tlog\link.write.1.tlog
