+++ PLT_Parser/PLT_Parser/x64/Debug/PLT_Parser.log
@@ -1,1 +1,9 @@
-﻿  PLT_Parser.vcxproj -> D:\Projects\c#\PLangTranslator\PLT_Parser\x64\Debug\PLT_Parser.exe
+﻿  Process "SyntaxAnal.y" bison file
+  Process "LexAnal.l" flex file
+  LexAlgorithms.c
+  LexAnal.flex.c
+  SyntaxAnal.tab.c
+SyntaxAnal.tab.c(1071,41): warning C4244: 'initializing': conversion from '__int64' to 'long', possible loss of data
+  Generating Code...
+  main.cpp
+  PLT_Parser.vcxproj -> D:\Projects\c#\PLangTranslator\PLT_Parser\x64\Debug\PLT_Parser.exe
