+++ PLT_Parser/PLT_Parser/x64/Debug/PLT_Parser.write.1.tlog
@@ -1,59 +1,3 @@
-^Parser.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Parser.tab.h
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Parser.tab.c
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Parser.tab.cpp
-^LexAnalyzer.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\LexAnalyzer.yy.c
-^Parser.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Parser.tab.h
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Parser.tab.c
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Parser.tab.cpp
-^LexAnalyzer.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\LexAnalyzer.yy.c
-^PyToAsmSyntax.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\PyToAsmSyntax.tab.h
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\PyToAsmSyntax.tab.c
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\PyToAsmSyntax.tab.cpp
-^Grammars\PyToAsm\SyntaxAnal.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\SyntaxAnal.tab.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\SyntaxAnal.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\SyntaxAnal.tab.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\SyntaxAnal.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\SyntaxAnal.tab.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\SyntaxAnal.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\SyntaxAnal.tab.cpp
-^Grammars\PyToAsm\SyntaxAnal.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\SyntaxAnal.tab.cpp
-^Grammars\PyToAsm\SyntaxAnal.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\SyntaxAnal.tab.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\SyntaxAnal.y
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\SyntaxAnal.tab.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
-^Grammars\PyToAsm\LexAnal.l
-D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\LexAnal.flex.cpp
 ^Grammars\PyToAsm\SyntaxAnal.y
 D:\Projects\c#\PLangTranslator\PLT_Parser\PLT_Parser\Grammars\PyToAsm\SyntaxAnal.tab.cpp
 ^Grammars\PyToAsm\LexAnal.l
